new gameplaytags

+++ StatSystem/StatSystem/Scripts/IDamage.cs
@@ -4,7 +4,7 @@ using UnityEngine;
 public interface IDamage
 { 
     bool IsCriticalHit { get; }
-    int Magnitude { get; }
+    float Magnitude { get; }
     GameObject Instigator { get; }
     object Source { get; }
 }
